Amend png image fallback code
The "onerror" handler for the SVG image that aims at replacing the image with a
PNG version created an infinite loop in the (unlikely, granted) event when the
client lost her network connection after having loaded the spec but before
having loaded the SVG image. See discussion in:
https://github.com/w3c/accelerometer/pull/55

This fix only changes the source URL once.