Rewrite Automation section after w3c/sensors#470
The Automation section in the Generic Sensor API specification was rewritten and several terms and concepts have changed.

This commit adapts the Magnetometer Sensor spec to the changes:

Remove references to "mock sensor type", "mock sensor reading values" and the "MockSensorType" enum.
Define an entry in the per-type virtual sensor metadata map whose key is what used to be the "magnetometer" or "uncalibrated-magnetometer" entry in MockSensorType and an appropriate virtual sensor metadata entry.
This is enough to integrate properly with the Generic Sensor spec and allow Magnetometer and Uncalibrated Magnetometer virtual sensors to be created and used.

Fixes #65.